cli/ts: Refaktoriser hjælpefunktioner til udtræk af tidsserier
