+++ fire/cli/ts/gnss.py
@@ -1,21 +1,12 @@
-from typing import Type
-from datetime import datetime
-
 import click
-import pandas as pd
 from pathlib import Path
 from pyproj import Transformer
-from rich.table import Table
-from rich.console import Console
-from rich import box
-from sqlalchemy import func
 from sqlalchemy.exc import NoResultFound
 
 import fire.cli
 from fire.api.model import (
     Tidsserie,
     GNSSTidsserie,
-    Punkt,
 )
 from fire.api.model.tidsserier import (
     TidsserieEnsemble,
@@ -29,52 +20,13 @@ from fire.cli.ts._plot_gnss import (
     GNSS_TS_PLOTTING_LABELS,
 )
 
+from fire.cli.ts import (
+    _find_tidsserie,
+    _print_tidsserieoversigt,
+    _udtræk_tidsserie,
+)
+
 from . import ts
-
-
-def _print_tidsserieoversigt(tidsserieklasse: Type, punkt: Punkt = None):
-    """
-    Oversigt over tidsserier af en bestemt types
-
-    raises:     SystemExit
-    """
-    if punkt:
-        tidsserier = [ts for ts in punkt.tidsserier if isinstance(ts, tidsserieklasse)]
-    else:
-        tidsserier = (
-            fire.cli.firedb.session.query(tidsserieklasse)
-            .filter(tidsserieklasse._registreringtil == None)
-            .all()
-        )  # NOQA
-
-    if not tidsserier:
-        raise SystemExit("Fandt ingen tidsserier")
-
-    tabel = Table("Ident", "Tidsserie ID", "Referenceramme", box=box.SIMPLE)
-
-    for ts in tidsserier:
-        tabel.add_row(ts.punkt.gnss_navn, ts.navn, ts.referenceramme)
-
-    console = Console()
-    console.print(tabel)
-
-
-def _find_tidsserie(tidsserieklasse: Type, tidsserienavn: str) -> Tidsserie:
-    """
-    Find en navngiven tidsserie
-
-    raises:     NoResultFound
-    """
-    tidsserie = (
-        fire.cli.firedb.session.query(tidsserieklasse)
-        .filter(
-            tidsserieklasse._registreringtil == None,
-            func.lower(tidsserieklasse.navn) == func.lower(tidsserienavn),
-        )
-        .one()
-    )  # NOQA
-
-    return tidsserie
 
 
 GNSS_TS_PARAMETRE = {
@@ -149,7 +101,7 @@ def gnss(objekt: str, parametre: str, fil: click.Path, **kwargs) -> None:
 
     "OBJEKT" sættes til enten et punkt eller et specifik navngiven tidsserie.
     Hvis "OBJEKT" er et punkt udskrives en oversigt over de tilgængelige
-    tidsserier til dette punkt. Hvis 'OBJEKT' er en tidsserie udskrives
+    tidsserier til dette punkt. Hvis "OBJEKT" er en tidsserie udskrives
     tidsserien på skærmen. Hvilke parametre der udskrives kan specificeres
     i en kommasepareret liste med ``--parametre``. Følgende parametre kan vælges::
 
@@ -206,62 +158,7 @@ def gnss(objekt: str, parametre: str, fil: click.Path, **kwargs) -> None:
 
         fire ts gnss RDIO_5D_IGb08 -p alle -f RDIO_5D_IGb08.xlsx
     """
-    if not objekt:
-        _print_tidsserieoversigt(GNSSTidsserie)
-        raise SystemExit
-
-    # Prøv først med at søg efter specifik tidsserie
-    try:
-        tidsserie = _find_tidsserie(GNSSTidsserie, objekt)
-    except NoResultFound:
-        try:
-            punkt = fire.cli.firedb.hent_punkt(objekt)
-        except NoResultFound:
-            raise SystemExit("Punkt eller tidsserie ikke fundet")
-
-        _print_tidsserieoversigt(GNSSTidsserie, punkt)
-        raise SystemExit
-
-    if parametre.lower() == "alle":
-        parametre = ",".join(GNSS_TS_PARAMETRE.keys())
-
-    parametre = parametre.split(",")
-    overskrifter = []
-    kolonner = []
-    for p in parametre:
-        if p not in GNSS_TS_PARAMETRE.keys():
-            raise SystemExit(f"Ukendt tidsserieparameter '{p}'")
-
-        overskrifter.append(p)
-        kolonner.append(tidsserie.__getattribute__(GNSS_TS_PARAMETRE[p]))
-
-    tabel = Table(*overskrifter, box=box.SIMPLE)
-    data = list(zip(*kolonner))
-
-    def klargør_celle(input):
-        if isinstance(input, datetime):
-            return str(input)
-        if isinstance(input, float):
-            return f"{input:.4f}"
-        if not input:
-            return ""
-
-    for række in data:
-        tabel.add_row(
-            *[klargør_celle(celle) if celle is not None else "" for celle in række]
-        )
-
-    console = Console()
-    console.print(tabel)
-
-    if not fil:
-        raise SystemExit
-
-    data = {
-        overskrift: kolonne for (overskrift, kolonne) in zip(overskrifter, kolonner)
-    }
-    df = pd.DataFrame(data)
-    df.to_excel(fil, index=False)
+    _udtræk_tidsserie(objekt, GNSSTidsserie, GNSS_TS_PARAMETRE, parametre, fil)
 
 
 @ts.command()
